added repayment processing , linked customer to collaterals

+++ src/main/java/com/sahay/loan/dto/CollateralDocumentDto.java
@@ -12,7 +12,7 @@ import org.springframework.web.multipart.MultipartFile;
 @NoArgsConstructor
 public class CollateralDocumentDto {
 
-    private Long collateralId;
+    private Integer collateralId;
     private MultipartFile document;
     private String description;
