+++ src/main/java/com/sahay/loan/repo/CollateralRepository.java
@@ -4,12 +4,14 @@ import com.sahay.loan.entity.Collateral;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface CollateralRepository extends JpaRepository<Collateral, Long> {
+public interface CollateralRepository extends JpaRepository<Collateral, Integer> {
 
     Optional<Collateral> findByCustomerId(int customerId);
+
     Optional<Collateral> findByCollateralNumber(String collateralNumber);
